refactor(gym/scenarios): split class_specific_checks into 6 helper modules
checks.rs (1741) → checks.rs (52) + checks_1..6.rs (~290-317 each).
Extracts 32 match arms into per-class helper fns, distributed across 6 files
via greedy bin packing. Dispatcher now a thin match table.

All 419 gym tests pass.

Refs #1266
